requirements.txt
comtypes==1.1.7
MouseInfo==0.1.3
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.1.2
PyAutoGUI==0.9.50
PyAutoIt==0.6.0
PyDirectInput==1.0.1
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pynput==1.6.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
pywin32==227
pywinauto==0.6.8
six==1.15.0
tflearn==0.3.2

image~=1.5.32
tesseract~=0.1.3
requests~=2.23.0
pytesseract~=0.3.4